pyproject.toml
[project]
name = "carting"
version = "0.1.0"
description = "A FastAPI-based RESTful API for cart management"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.11",
    "sqlalchemy>=2.0.39",
    "uvicorn>=0.34.0",
    "python-dotenv>=1.0.1",
    "pydantic>=2.7.4",
    "pydantic-settings>=2.4.0",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.15.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.1",
    "httpx>=0.27.0",
    "black>=24.2.0",
    "isort>=5.13.2",
    "flake8>=7.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
